refactor: add size util
