+++ paddle/fluid/pybind/process_group_utils.h
@@ -21,7 +21,7 @@
 #include "paddle/phi/kernels/funcs/concat_and_split_functor.h"
 
 namespace paddle {
-namespace distributed {
+namespace pybind {
 
 template <typename DeviceContext, typename T>
 struct ConcatDenseTensor {
@@ -257,5 +257,10 @@ void SplitTensor(const phi::DeviceContext &dev_ctx,
   }
 }
 
-}  //  namespace distributed
+inline std::vector<int64_t> GetDefaultSplitSizes(const phi::DenseTensor &tensor,
+                                                 int world_size) {
+  return std::vector<int64_t>(world_size, tensor.dims()[0] / world_size);
+}
+
+}  //  namespace pybind
 }  //  namespace paddle
